+++ src/plone/restapi/tests/test_services_roles.py
@@ -1,8 +1,10 @@
+from plone import api
 from plone.app.testing import SITE_OWNER_NAME
 from plone.app.testing import SITE_OWNER_PASSWORD
 from plone.restapi.testing import PLONE_RESTAPI_DX_FUNCTIONAL_TESTING
 from plone.restapi.testing import RelativeSession
 
+import transaction
 import unittest
 
 
@@ -21,7 +23,74 @@ class TestRolesGet(unittest.TestCase):
     def tearDown(self):
         self.api_session.close()
 
+    def set_siteadm(self):
+        siteadm_username = "siteadm"
+        siteadm_password = "siteadmpassword"
+        api.user.create(
+            email="siteadm@example.com",
+            roles=["Site Administrator"],
+            username=siteadm_username,
+            password=siteadm_password,
+        )
+        self.api_session = RelativeSession(self.portal_url, test=self)
+        self.api_session.headers.update({"Accept": "application/json"})
+        self.api_session.auth = (siteadm_username, siteadm_password)
+        transaction.commit()
+
     def test_roles_endpoint_lists_roles(self):
+        response = self.api_session.get("/@roles")
+
+        expected = (
+            {
+                "@id": self.portal_url + "/@roles/Contributor",
+                "@type": "role",
+                "id": "Contributor",
+                "title": "Contributor",
+            },
+            {
+                "@id": self.portal_url + "/@roles/Editor",
+                "@type": "role",
+                "id": "Editor",
+                "title": "Editor",
+            },
+            {
+                "@id": self.portal_url + "/@roles/Member",
+                "@type": "role",
+                "id": "Member",
+                "title": "Member",
+            },
+            {
+                "@id": self.portal_url + "/@roles/Reader",
+                "@type": "role",
+                "id": "Reader",
+                "title": "Reader",
+            },
+            {
+                "@id": self.portal_url + "/@roles/Reviewer",
+                "@type": "role",
+                "id": "Reviewer",
+                "title": "Reviewer",
+            },
+            {
+                "@id": self.portal_url + "/@roles/Site Administrator",
+                "@type": "role",
+                "id": "Site Administrator",
+                "title": "Site Administrator",
+            },
+            {
+                "@id": self.portal_url + "/@roles/Manager",
+                "@type": "role",
+                "id": "Manager",
+                "title": "Manager",
+            },
+        )
+        result = response.json()
+        self.assertEqual(len(expected), len(result))
+        for item in result:
+            self.assertIn(item, expected)
+
+    def test_siteadm_roles_endpoint_lists_roles(self):
+        self.set_siteadm()
         response = self.api_session.get("/@roles")
 
         expected = (